COTR-66 Implemented basic loot drop functionality
