+++ Scenes/Actors/Troops/Infantry.cs
@@ -31,4 +31,18 @@ public partial class Infantry : Troop
     /// Method for playing attacking animation
     /// </summary>
     public virtual void PlayAttacking() { }
+
+    /// <summary>
+    /// Method for handling received damage
+    /// </summary>
+    /// <param name="baseDamage">amount of received damage</param>
+    /// <param name="impactPosition">position for calculating particles casting direciton</param>
+    public override void HandleHit(float baseDamage, Vector2 impactPosition)
+    {
+        Blood blood = bloodScene.Instantiate() as Blood;
+        GetParent().AddChild(blood);
+        blood.GlobalPosition = GlobalPosition;
+        blood.Rotation = impactPosition.DirectionTo(GlobalPosition).Angle();
+        base.HandleHit(baseDamage, impactPosition);
+    }
 }
