+++ Scenes/Actors/Troops/Infantry/Footman.cs
@@ -136,20 +136,6 @@ public partial class Footman : Infantry
     }
 
     /// <summary>
-    /// Method for handling received damage
-    /// </summary>
-    /// <param name="baseDamage">amount of received damage</param>
-    /// <param name="impactPosition">position for calculating particles casting direciton</param>
-    public override void HandleHit(float baseDamage, Vector2 impactPosition)
-    {
-        base.HandleHit(baseDamage, impactPosition);
-        Blood blood = bloodScene.Instantiate() as Blood;
-        GetParent().AddChild(blood);
-        blood.GlobalPosition = GlobalPosition;
-        blood.Rotation = impactPosition.DirectionTo(GlobalPosition).Angle();
-    }
-
-    /// <summary>
     /// Method for handling attack timer timeout
     /// </summary>
     private void AttackTimerTimeout()
