+++ Scenes/Actors/Troops/Troop.cs
@@ -1,23 +1,29 @@
 using Godot;
+using System;
 
 public partial class Troop : Actor
 {
     [Signal] public delegate void TroopDiedEventHandler();
+    protected Globals globals; // global variables and functionality
 
     // Called when the node enters the scene tree for the first time.
     public override void _Ready()
     {
         base._Ready();
+        globals = GetNode<Globals>("/root/Globals");
     }
 
     /// <summary>
-    /// Remove health from the troop
+    /// Method for handling received damage
     /// </summary>
-    /// <param name="baseDamage">Ammount of health to remove</param>
-    /// <param name="impactPosition">Position of the hit</param>
+    /// <param name="baseDamage">amount of received damage</param>
+    /// <param name="impactPosition">position for calculating particles casting direciton</param>
     public override void HandleHit(float baseDamage, Vector2 impactPosition)
     {
-        base.HandleHit(baseDamage, impactPosition);
+        if (Stats.Health <= 0)
+        {
+            Die();
+        }
     }
 
     /// <summary>
@@ -27,6 +33,8 @@ public partial class Troop : Actor
     {
         if (Stats.Gold > 0)
         {
+            Random rand = new Random();
+            globals.EmitSignal("CoinsDroped", rand.Next(1, Stats.Gold), GlobalPosition);
         }
         EmitSignal(nameof(TroopDied));
         base.Die();
